Split Index and Status into separate files
