+++ kirin/resources/status.py
@@ -30,20 +30,11 @@
 # www.navitia.io
 
 from __future__ import absolute_import, print_function, unicode_literals, division
-from flask_restful import Resource, url_for
+from flask_restful import Resource
 import kirin
 from kirin.version import version
 from flask import current_app
 from kirin.core import model
-
-
-class Index(Resource):
-    def get(self):
-        response = {
-            "status": {"href": url_for("status", _external=True)},
-            "cots": {"href": url_for("cots", _external=True)},
-        }
-        return response, 200
 
 
 class Status(Resource):
